Add files via upload
it is a linked list with header node.
header node is the dummy node that is present at the begining of the list and its link part is used to store the address of the first actual node  of the list. the info part of this node may be empty or can be used to contain useful information about the list like  count of elements currently present in the list.